예외처리 1차 처리

+++ src/components/CommentList.jsx
@@ -4,7 +4,9 @@ import { useDispatch } from "react-redux";
 const CommentList = (props) => {
   const dispatch = useDispatch();
   const onDeleteButton = (payload) => {
-    dispatch(props.__deleteComment(payload));
+    if (window.confirm("정말 삭제하시겠습니까?")) {
+      dispatch(props.__deleteComment(payload));
+    }
   };
 
   return (
